pyproject.toml
[project]
name = "camera_module"
version = "3.0.0"
description = "A simple MADSci node module that supports taking snapshots with a camera or other video device."
authors = [
    {name = "Ryan D. Lewis", email="ryan.lewis@anl.gov"},
    {name = "Tobias Ginsburg", email="tginsburg@anl.gov"}
]
dependencies = [
    "madsci.node_module~=0.6.0",
    "madsci.client~=0.6.0",
    "madsci.common~=0.6.0",
]
requires-python = ">=3.9.1"
readme = "README.md"
license = {text = "MIT"}

[project.urls]
homepage = "https://github.com/AD-SDL/camera_module"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "ipykernel>=6.31.0",
]
